+++ example/lib/injector.dart
@@ -1,4 +1,3 @@
-import 'package:example/register_module.dart';
 import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
 
@@ -9,18 +8,6 @@ final getIt = GetIt.instance;
 @injectableInit
 Future<void> configure() async {
   $initGetIt(getIt);
-
-  getIt.registerSingletonWithDependencies(() => ServiceAA(),
-      dependsOn: [ServiceX]);
-
-  // getIt.registerFactory(() => ServiceAA());
-  // getIt.registerFactoryAsync(() => ServiceAA.createService());
-  // getIt.registerSingletonAsync(() => ServiceAA.createService());
-  // getIt.registerSingletonAsync(() => ServiceAA.createService());
-  // getIt.registerSingletonWithDependencies(() => ServiceAA.createService());
-  // getIt.registerLazySingletonAsync<SharedPreferences>(
-  //     () async => SharedPreferences.getInstance());
-
-  // getIt.registerFactoryParam<Dio, String, void>(
-  //     (s, _) => Dio(BaseOptions(baseUrl: s)));
+  // getIt.registerFactoryParam<ProductService, String, dynamic>(
+  //     (s, d) => ProductService(s));
 }