+++ injectable_generator/lib/generator/factory_generator.dart
@@ -1,41 +1,21 @@
 import 'package:injectable_generator/dependency_config.dart';
 import 'package:injectable_generator/generator/register_func_generator.dart';
 
-class LazyFactoryGenerator extends RegisterFuncGenerator {
+class FactoryGenerator extends RegisterFuncGenerator {
   @override
   String generate(DependencyConfig dep) {
     final constructBody = dep.moduleConfig == null
         ? generateConstructor(dep)
         : generateConstructorForModule(dep);
 
-    final asyncStr = dep.isAsync && !dep.asInstance ? 'Async' : '';
-    writeln("g.registerFactory$asyncStr<${dep.type}>(()=> $constructBody");
-    if (dep.instanceName != null) {
-      write(",instanceName: '${dep.instanceName}'");
+    var constructor = constructBody;
+    if (dep.registerAsInstance) {
+      constructor = generateAwaitSetup(dep, constructBody);
     }
-    write(");");
-    return buffer.toString();
-  }
-}
 
-class LazySingletonGenerator extends LazyFactoryGenerator {
-  @override
-  String generate(DependencyConfig dep) {
-    final constructBody = dep.moduleConfig == null
-        ? generateConstructor(dep)
-        : generateConstructorForModule(dep);
-
-    final asyncStr = dep.isAsync && !dep.asInstance ? 'Async' : '';
-    writeln(
-        "g.registerLazySingleton$asyncStr<${dep.type}>(()=> $constructBody");
-
-    if (dep.signalsReady != null) {
-      write(',signalsReady: ${dep.signalsReady}');
-    }
-    if (dep.instanceName != null) {
-      write(",instanceName: '${dep.instanceName}'");
-    }
-    write(");");
+    final asyncStr = dep.isAsync && !dep.preResolve ? 'Async' : '';
+    writeln("g.registerFactory$asyncStr<${dep.type}>(()=> $constructor");
+    closeRegisterFunc(dep);
     return buffer.toString();
   }
 }